splitting cli & lib
